ioctl cmd arg is u_long
SCCS-vsn: sys/netinet/in_var.h 8.2
SCCS-vsn: sys/netinet/in.c 8.4
SCCS-vsn: sys/netinet/tcp_usrreq.c 8.3
SCCS-vsn: sys/netinet/ip_var.h 8.2
SCCS-vsn: sys/netinet/udp_usrreq.c 8.5

+++ usr/src/sys/netinet/in.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)in.c	8.3 (Berkeley) %G%
+ *	@(#)in.c	8.4 (Berkeley) %G%
  */
 
 #include <sys/param.h>
@@ -132,7 +132,7 @@ extern	struct ifnet loif;
 /* ARGSUSED */
 in_control(so, cmd, data, ifp)
 	struct socket *so;
-	int cmd;
+	u_long cmd;
 	caddr_t data;
 	register struct ifnet *ifp;
 {
